Create gzip in memory

+++ internal/storage/writer-gzip.go
@@ -1,7 +1,9 @@
 package storage
 
 import (
+	"bytes"
 	"compress/gzip"
+	"io"
 	"os"
 
 	"github.com/c8121/asset-storage/internal/config"
@@ -10,9 +12,13 @@ import (
 type StorageZipFileWriter struct {
 	File      *os.File
 	zipWriter *gzip.Writer
+	outWriter io.Writer
 }
 
 func (writer *StorageZipFileWriter) Name() string {
+	if writer.File == nil {
+		return "InMemoryZip"
+	}
 	return writer.File.Name()
 }
 
@@ -27,15 +33,33 @@ func (writer *StorageZipFileWriter) Close() error {
 	if err := writer.zipWriter.Close(); err != nil {
 		return err
 	}
-	return writer.File.Close()
+	if writer.File != nil {
+		return writer.File.Close()
+	}
+	return nil
 }
 
 func (writer *StorageZipFileWriter) Move(path string) error {
-	return os.Rename(writer.Name(), path)
+	if writer.File != nil {
+		return os.Rename(writer.Name(), path)
+	} else {
+		out, err := os.Create(path)
+		if err != nil {
+			return err
+		}
+		defer out.Close()
+
+		buf := writer.outWriter.(*bytes.Buffer)
+		_, err = out.Write(buf.Bytes())
+		return err
+	}
 }
 
 func (writer *StorageZipFileWriter) Remove() error {
-	return os.Remove(writer.Name())
+	if writer.File != nil {
+		return os.Remove(writer.Name())
+	}
+	return nil
 }
 
 func NewTempZipFileWriter() (*StorageZipFileWriter, error) {
@@ -46,7 +70,20 @@ func NewTempZipFileWriter() (*StorageZipFileWriter, error) {
 
 	zip := &StorageZipFileWriter{}
 	zip.File = file
-	zip.zipWriter = gzip.NewWriter(file)
+	zip.outWriter = file
+	zip.zipWriter = gzip.NewWriter(zip.outWriter)
+
+	return zip, nil
+}
+
+func NewMemZipFileWriter() (*StorageZipFileWriter, error) {
+
+	buf := bytes.NewBuffer(make([]byte, 0, config.MaxMemFileSize))
+
+	zip := &StorageZipFileWriter{}
+	zip.File = nil
+	zip.outWriter = buf
+	zip.zipWriter = gzip.NewWriter(zip.outWriter)
 
 	return zip, nil
 }
